Move save_mrc_tofile to MasterRecessionCalcTool
